# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hawkeye Sterling API",
    "version": "0.2.0",
    "description": "Regulator-grade AML/CFT/sanctions/PEP/adverse-media screening engine. Stateless, explainable, audit-chained.",
    "contact": {
      "name": "Hawkeye Sterling support",
      "url": "https://hawkeye-sterling.netlify.app/status"
    },
    "license": { "name": "Proprietary" }
  },
  "servers": [
    { "url": "https://hawkeye-sterling.netlify.app", "description": "Production" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/schedule": {
      "get": { "summary": "List per-subject rescreening schedules", "tags": ["Monitoring"], "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Create or update a per-subject schedule",
        "tags": ["Monitoring"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["subjectId", "cadence"],
                "properties": {
                  "subjectId": { "type": "string" },
                  "cadence": { "type": "string", "enum": ["hourly", "daily", "weekly", "monthly"] },
                  "scoreThreshold": { "type": "number", "minimum": 0, "maximum": 1 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Schedule persisted" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key (hks_live_…)"
      }
    },
...
  }
}